Create tính chi phí gửi thư
All cities in Lineland lie on the Ox coordinate axis. Therefore, each city is associated with the position XI the coordinate on the Ox axis. No two cities are located at the same point. Lineland residents like to send letters to each other. A person can send a letter only if the recipient lives in another city. The cost of sending a letter is exactly equal to the distance between the sender's city and the recipient's city. For each city, calculate two values mini and maxi, where Min_I is the minimum cost of sending a letter from the I-th city to another city and Max_I is the maximum cost of sending a letter from the i-th city to some other city