Create bán vé và trả tiền
The movie "Die Hard" has just been released! There are N people at the cinema theater in a large row.
Each of them has a banknote in denominations of 100, 50 or 25 dollars. A "Die Hard" ticket costs 25 dollar.
Can the reservationist sell tickets to each person and pay a lot if he initially has no money and sells the tickets in the correct order to everyone in the line?